// plugins/plugins-types.ts
export type BuiltinsWithOptionalParams = DefaultPlugins & {
  'preset-default': {
    floatPrecision?: number
    /**
     * All default plugins can be customized or disabled here
     * for example
     * {
     *   sortAttrs: { xmlnsOrder: "alphabetical" },
     *   cleanupAttrs: false,
     * }
     */
    overrides?: PresetDefaultOverrides
  }
  cleanupListOfValues: {
    floatPrecision?: number
    leadingZero?: boolean
    defaultPx?: boolean
    convertToPx?: boolean
  }
  convertStyleToAttrs: {
    keepImportant?: boolean
  }
  prefixIds: {
    prefix?:
      | boolean
      | string
      | ((node: XastElement, info: PluginInfo) => string)
    delim?: string
    prefixIds?: boolean
    prefixClassNames?: boolean
  }
  removeDimensions: void
  removeOffCanvasPaths: void
  removeRasterImages: void
  removeScriptElement: void
  removeStyleElement: void
  removeXMLNS: void
  reusePaths: void
}

export type BuiltinsWithRequiredParams = {
  addAttributesToSVGElement: {
    attribute?: string | Record<string, null | string>
    attributes?: Array<string | Record<string, null | string>>
  }
  addClassesToSVGElement: {
    className?: string
    classNames?: Array<string>
  }
  removeAttributesBySelector: any
  removeAttrs: {
    elemSeparator?: string
    preserveCurrentColor?: boolean
    attrs: string | Array<string>
  }
  removeElementsByAttr: {
    id?: string | Array<string>
    class?: string | Array<string>
  }
}

type PluginsParams = BuiltinsWithOptionalParams & BuiltinsWithRequiredParams

export type Plugin<Name extends keyof PluginsParams> = PluginDef<
  PluginsParams[Name]
>